LPS-88392 Move test dependency files out of the fake bundle packages
