+++ portal-impl/test/integration/com/liferay/portal/kernel/portlet/bridges/mvc/MVCActionCommandTest.java
@@ -18,9 +18,9 @@ import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.model.Portlet;
 import com.liferay.portal.kernel.portlet.LiferayPortletConfig;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
-import com.liferay.portal.kernel.portlet.bridges.mvc.bundle.mvcactioncommand.TestMVCActionCommand1;
-import com.liferay.portal.kernel.portlet.bridges.mvc.bundle.mvcactioncommand.TestMVCActionCommand2;
-import com.liferay.portal.kernel.portlet.bridges.mvc.bundle.mvcactioncommand.TestPortlet;
+import com.liferay.portal.kernel.portlet.bridges.mvc.mvcactioncommand.TestMVCActionCommand1;
+import com.liferay.portal.kernel.portlet.bridges.mvc.mvcactioncommand.TestMVCActionCommand2;
+import com.liferay.portal.kernel.portlet.bridges.mvc.mvcactioncommand.TestPortlet;
 import com.liferay.portal.kernel.util.JavaConstants;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
