--- portal-impl/test/integration/com/liferay/portal/kernel/portlet/bridges/mvc/bundle/mvcactioncommand/TestPortlet.java
@@ -1,28 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.kernel.portlet.bridges.mvc.bundle.mvcactioncommand;
-
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
-
-/**
- * @author Manuel de la Peña
- */
-public class TestPortlet extends MVCPortlet {
-
-	public static final String PORTLET_NAME =
-		"com_liferay_portal_kernel_portlet_bridges_mvc_bundle_actioncommand_" +
-			"TestPortlet";
-
-}
